% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_surnames.R
\name{clean_surnames}
\alias{clean_surnames}
\title{Clean a vector of (English-language) surnames}
\usage{
clean_surnames(raw_names)
}
\arguments{
\item{raw_names}{Character vector of raw surname strings to be cleaned.}
}
\value{
A character vector of cleaned or blank names the same length as `raw_names`
}
\description{
`clean_surnames()` takes a character vector of raw surname strings and
returns a vector of "cleaned" or empty strings if the input is too ambiguous.
The rules are designed for handling surnames in historical US census data,
but may work well for English-language surnames more broadly.
}
\details{
The function performs the following steps:
\enumerate{
  \item Convert all strings to lowercase.
  \item Remove common suffixes “jr” and “sr”.
  \item Strip spaces and apostrophes.
  \item Remove all digits.
  \item Standardize “missing‐letter” markers:
  \item Blank out remaining names still containing non-letter/space characters.
  \item Blank out known placeholders: “unreadable”, “unknown”, etc.
  \item Blank out any name of length ≤1.
  \item Blank out uncommon names of length 2.
  \item Blank out names with >25% missing characters.
}
}
