% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicknamer.R
\name{names}
\alias{names}
\alias{nicknamer}
\title{Get citations for CSPP variables}
\usage{
nicknamer(
  format = "long",
  transform = NULL,
  reverse = FALSE,
  opt_function = NULL
)
}
\arguments{
\item{format}{Default is long. Takes either "long" or "wide". The long
version contains one row per name-nickname match. The wide format contains
one row per name, with all possible nicknames in individual columns.}

\item{transform}{Default is NULL, resulting in no string transformation to
names. Options are "to_upper" or "to_lower", resulting in lower case or
uppercase transformations respectively.}

\item{reverse}{Default is FALSE. If set to true, the nickname and names are
switched. This will result in different output when the format is set to
"wide" as it will now be one row per nickname as opposed to one row per
name. E.g.: "Alex" as a nickname will now be associated with "Alexander"
and "Alexandra".}

\item{opt_function}{Default is NULL. Allows the user to pass custom functions
to the name and nickname columns, such as \code{stringr::str_sub} or
any user-written function that parses strings.}
}
\value{
dataframe
}
\description{
\code{nicknamer} generates a dataframe of common nicknames matched to first
names. The resulting dataframe can be in wide or long format.
}
\examples{

# standard long dataframe
names <- nicknamer()

# wide dataframe transformed to lower case
names.wide <- nicknamer(format="wide", transform = "to_lower")

# wide dataframe with nickname as the row
names.wide <- nicknamer(format="wide", transform = "to_lower", reverse=T)

# optional function applied to names:
short.names <- nicknamer(
  format="wide",
  transform = "to_lower",
  opt_function = function(x) { str_sub(x, 1, 3) }
)
}
