% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicknames.R
\docType{data}
\name{nicknames}
\alias{nicknames}
\title{Nicknames Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1431 rows and 2 columns.
}
\usage{
nicknames
}
\description{
@docType data
}
\details{
@usage data(nicknames)

 @format The primary dataset containing names and common nickname variations.

 @keywords datasets
}
\keyword{datasets}
