% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdirichlet.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Generate a random draw from a Dirichlet distribution}
\usage{
rdirichlet(alpha)
}
\arguments{
\item{alpha}{Numeric vector; concentration parameters of the Dirichlet (length K).}
}
\value{
Numeric vector of length K that sums to 1, a single Dirichlet(α) draw.
}
\description{
Draws a single random vector from a Dirichlet distribution using the gamma trick.
}
\examples{
rdirichlet(c(1, 1, 1))  # uniform over 3-simplex

}
