% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_names.R
\name{standardize_names}
\alias{standardize_names}
\title{Return Bayes choices for standard names}
\usage{
standardize_names(names, dictionary, lambda = 1, post, ncores = 1)
}
\arguments{
\item{names}{A vector of character strings to standardize. It's assumed that
name strings have already been "cleaned" in the same manner as
input to `make_bayes_choice_dictionary()`, e.g. using
`clean_surnames()` for English-language surnames.}

\item{dictionary}{The output of `make_bayes_choice_dictionary()`.
data.frame with two columns:
       $observed: character vector of observed, noisy names
       $standard: character vector of standardized names}

\item{lambda}{Should match the `data` argument in `draw_gibbs()`.
Numeric; fixed scale parameter for error kernel (default 1)}

\item{post}{The output of `draw_gibbs()`, a list with three elementS:
- `delta_samples` the chain of draws for \eqn{\delta}
- `x_avg` the posterior mean for the probability of inclusion.
- `p_avg` the posterior mean frequency in the sample.}

\item{ncores}{Integer ≥ 1; number of parallel worker processes to launch.
Defaults to 1.}
}
\value{
A data.frame with two columns:
  - `observed`: string vector of observed names.
  - `standard`: string vector of standardized names.
}
\description{
takes a vector of raw name strings and returns standardized names.
Relies on the output of `draw_gibbs()` and `make_bayes_choice_dictionary`.
Bayes' choice for a name \eqn{s_i} is \eqn{s_j} to maximize \eqn{Pr(\text{true surname} = s_j | \text{observed} = s_i, \lambda, \delta, s_j \in C)}
where \eqn{C} collects names included more than 99% of the time as true names in posterior draws.
}
