% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_us_surnames.R
\name{standardize_us_surnames}
\alias{standardize_us_surnames}
\title{Return Bayes choices for surnames in US historical censuses}
\usage{
standardize_us_surnames(names)
}
\arguments{
\item{names}{A vector of character strings to standardize. It's assumed that
name strings have already been "cleaned" e.g. using
`clean_surnames()` for English-language surnames.}
}
\value{
A data.frame with two columns:
  - `observed`: string vector of observed names.
  - `standard`: string vector of standardized names.
}
\description{
takes a vector of cleaned name strings and returns standardized names.
Jumps to the final dictionary for
}
