% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_us_dictionary.R
\name{load_us_dictionary}
\alias{load_us_dictionary}
\title{Download, load, and delete a large list of "standardized" surnames from US censuses}
\usage{
load_us_dictionary()
}
\value{
A data frame containing a dictionary of observed and "standard" surnames.
}
\description{
Download, load, and delete a large list of "standardized" surnames from US censuses
}
