% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bayes_choice_dictionary.R
\name{make_bayes_choice_dictionary}
\alias{make_bayes_choice_dictionary}
\title{Build a dictionary of Bayes choices for standard choices from posterior draws}
\usage{
make_bayes_choice_dictionary(data, neighbor_list, lambda, post, ncores = 1)
}
\arguments{
\item{data}{Should match the \code{data} argument in \code{draw_gibbs()}.
Data frame with columns:
- string: unique identifiers
- count: integer counts}

\item{neighbor_list}{Should match the \code{data} argument in \code{draw_gibbs()}.
List of length K with elements:
$j: integer vector of neighbor indices
$d: numeric vector of edit distances}

\item{lambda}{Should match the \code{data} argument in \code{draw_gibbs()}.
Numeric; fixed scale parameter for error kernel (default 1)}

\item{post}{The output of \code{draw_gibbs()}, a list with three elementS:
- \code{delta_samples} the chain of draws for \eqn{\delta}
- \code{x_avg} the posterior mean for the probability of inclusion.
- \code{p_avg} the posterior mean frequency in the sample.}

\item{ncores}{Integer ≥ 1; number of parallel worker processes to launch.
Defaults to 1.}
}
\value{
A data.frame with two columns:
\itemize{
\item \code{observed}: string vector of observed names.
\item \code{standard}: string vector of standardized names (\code{NA} if name is ambiguous).
}
}
\description{
\code{make_bayes_choice_dictionary()} takes the input and output of \code{draw_gibbs()}
and returns a df with classified names. Bayes' choice for a name \eqn{s_i} is
\eqn{s_j} to maximize \eqn{Pr(\text{true surname} = s_j | \text{observed} = s_i, \lambda, \delta, s_j \in C)}
where \eqn{C} collects names included more than 99\% of the time as true names in posterior draws.
}
