% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bayes_choice_dictionary.R
\name{make_bayes_choice_dictionary}
\alias{make_bayes_choice_dictionary}
\title{Build a dictionary of Bayes choices for standard choices from posterior draws}
\usage{
make_bayes_choice_dictionary(data, D, lambda, post, ncores = 1)
}
\arguments{
\item{data}{Should match the `data` argument in `draw_gibbs()`.
Data frame with columns:
        - name: unique identifiers
        - count: integer counts}

\item{D}{Matrix or sparse matrix matching the argument to `draw_gibbs`; K×K distance matrix with distances for neighbor pairs.}

\item{post}{The output of `draw_gibbs()`, a list including three elements used here:
- `delta` the chain of draws for \eqn{\delta}
- `lambda` the chain of draws for \eqn{\lambda}
- `p_mean` the posterior mean for the latent frequency of names in the population.}

\item{ncores}{Integer ≥ 1; number of parallel worker processes to launch.
Defaults to 1.}
}
\value{
A data.frame with five columns:
   $observed: character vector of observed, noisy names
   $standard: character vector of standardized names
   $posterior: numeric vector of posterior probabilities
   $bayes_choice: integer indicating the posterior mode
   $p_standard: numeric vector containing frequency of the standard name
}
\description{
`make_bayes_choice_dictionary()` takes the input and output of `draw_gibbs()`
and returns a df with classified names. Bayes' choice for a name \eqn{s_i} is
\eqn{s_j} to maximize \eqn{Pr(\text{true name} = s_j | \text{observed} = s_i, \lambda, \delta, s_j \in C)}
where \eqn{C} collects names included more than 99% of the time as true names in posterior draws.
}
