% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bayes_choice_dictionary_helper.R
\name{make_bayes_choice_dictionary_helper}
\alias{make_bayes_choice_dictionary_helper}
\title{Helper function for `make_bayes_choice_dictionary()`}
\usage{
make_bayes_choice_dictionary_helper(pack)
}
\arguments{
\item{pack}{List of elements prepared by}
}
\value{
A data.frame with the `observed` name, id of the `standard` name,
       `posterior` probability of the classification, and `bayes_choice`
       indicator of the posterior mode.
}
\description{
Helper function for `make_bayes_choice_dictionary()`
}
