% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_missing_name.R
\name{standardize_missing_name}
\alias{standardize_missing_name}
\title{Helper function for `standardize_names()`}
\usage{
standardize_missing_name(name)
}
\arguments{
\item{name}{A character string to standardize.}
}
\value{
A character string of the standardized name or NA.
}
\description{
takes a vector of clean name strings, outside the dictionary!,
and returns standardized names. Relies on the output of `draw_gibbs()` and `make_bayes_choice_dictionary`.
}
