% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_names.R
\name{standardize_names}
\alias{standardize_names}
\title{Return Bayes choices for standard names}
\usage{
standardize_names(names, dictionary, lambda, delta, method = "jw", ncores = 1)
}
\arguments{
\item{names}{A vector of character strings to standardize. It's assumed that
name strings have already been "cleaned" in the same manner as
input to `make_bayes_choice_dictionary()`, e.g. using
`clean_surnames()` for English-language surnames.}

\item{dictionary}{The output of `make_bayes_choice_dictionary()`.
data.frame with two columns:
       $observed: character vector of observed, noisy names
       $standard: character vector of standardized names
       $p_standard: positive numeric vector of name frequencies}

\item{lambda}{Postive number equal to the posterior mean; i.e.`mean(post$lambda)`}

\item{delta}{Postive number equal to the posterior mean; i.e.`mean(post$delta)`
from the output of `draw_gibbs()`.}

\item{method}{method   Character scalar; distance metric passed to
`stringdist::stringdist()`.  Common choices: `"jw"`, `"lv"`
should match the method used in the original dictionary.
Default is "jw".}

\item{ncores}{Integer ≥ 1; number of parallel worker processes to launch.
Defaults to 1.}
}
\value{
A data.frame with two columns:
  - `observed`: string vector of observed names.
  - `standard`: string vector of standardized names.
}
\description{
takes a vector of cleaned name strings and returns standardized names.
Relies on the output of `draw_gibbs()` and `make_bayes_choice_dictionary`.
Bayes' choice for a name \eqn{s_i} is \eqn{s_j} to maximize \eqn{Pr(\text{true surname} = s_j | \text{observed} = s_i, \lambda, \delta, s_j \in C)}
where \eqn{C} collects names included more than 99% of the time as true names in posterior draws.
}
