% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_names.R,
%   R/standardize_us_surnames.R
\name{standardize_names}
\alias{standardize_names}
\title{Return Bayes' choices for standard names}
\usage{
standardize_names(data, neighbor_list, lambda, post, ncores = 1)

standardize_names(data, neighbor_list, lambda, post, ncores = 1)
}
\arguments{
\item{ncores}{Integer ≥ 1; number of parallel worker processes to launch.
Defaults to 1.}

\item{names}{A vector of unique character strings to standardize.
Data frame with columns:
- string: unique identifiers
- count: integer counts}

\item{dictionary}{The output of \code{make_bayes_choice_dictionary()}.
data.frame with two columns:
$observed: character vector of observed, noisy names
$standard: character vector of standardized names}
}
\value{
A data.frame with two columns:
\itemize{
\item \code{observed}: string vector of observed names.
\item \code{standard}: string vector of standardized names.
}

A data.frame with two columns:
\itemize{
\item \code{observed}: string vector of observed names.
\item \code{standard}: string vector of standardized names.
}
}
\description{
#' @description

#' @description
}
