% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_us_surnames.R
\name{standardize_us_surnames}
\alias{standardize_us_surnames}
\title{Return Bayes' choices for standard names}
\usage{
standardize_us_surnames(data, neighbor_list, lambda, post, ncores = 1)
}
\arguments{
\item{ncores}{Integer ≥ 1; number of parallel worker processes to launch.
Defaults to 1.}

\item{names}{A vector of unique character strings to standardize.
Data frame with columns:
        - string: unique identifiers
        - count: integer counts}

\item{dictionary}{The output of `make_bayes_choice_dictionary()`.
data.frame with two columns:
       $observed: character vector of observed, noisy names
       $standard: character vector of standardized names}
}
\value{
A data.frame with two columns:
  - `observed`: string vector of observed names.
  - `standard`: string vector of standardized names.
}
\description{
#' @description
}
